# frontend/lib/deskly-openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "TicketCreate": {
        "properties": {
          "asignado_a": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asignado A"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "prioridad": {
            "default": "medium",
            "maxLength": 32,
            "title": "Prioridad",
            "type": "string"
          },
          "titulo": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "titulo"
        ],
        "title": "TicketCreate",
        "type": "object"
      },
      "TicketRead": {
        "properties": {
          "actualizado_en": {
            "format": "date-time",
            "title": "Actualizado En",
            "type": "string"
          },
          "asignado_a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asignado A"
          },
          "creado_en": {
            "format": "date-time",
            "title": "Creado En",
            "type": "string"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "estado": {
            "title": "Estado",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "prioridad": {
            "title": "Prioridad",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "titulo",
          "descripcion",
          "prioridad",
          "estado",
          "asignado_a",
          "creado_en",
          "actualizado_en"
        ],
        "title": "TicketRead",
        "type": "object"
      },
      "TicketSortBy": {
        "enum": [
          "id",
          "creado_en",
          "actualizado_en",
          "prioridad",
          "estado",
          "asignado_a"
        ],
        "title": "TicketSortBy",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookTicketIngestRequest": {
        "properties": {
          "event_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Event Id",
            "type": "string"
          },
          "ticket": {
            "$ref": "#/components/schemas/WebhookTicketPayload"
          }
        },
        "required": [
          "event_id",
          "ticket"
        ],
        "title": "WebhookTicketIngestRequest",
        "type": "object"
      },
      "WebhookTicketPayload": {
        "properties": {
          "asignado_a": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asignado A"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "prioridad": {
            "default": "medium",
            "maxLength": 32,
            "title": "Prioridad",
            "type": "string"
          },
          "titulo": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Titulo",
            "type": "string"
          }
        },
        "required": [
          "titulo"
        ],
        "title": "WebhookTicketPayload",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Deskly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tickets": {
      "get": {
        "operationId": "list_tickets_api_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "in": "query",
            "name": "prioridad",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prioridad"
            }
          },
          {
            "in": "query",
            "name": "asignado_a",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asignado A"
            }
          },
          {
            "in": "query",
            "name": "creado_desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creado Desde"
            }
          },
          {
            "in": "query",
            "name": "creado_hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creado Hasta"
            }
          },
          {
            "in": "query",
            "name": "actualizado_desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actualizado Desde"
            }
          },
          {
            "in": "query",
            "name": "actualizado_hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actualizado Hasta"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TicketSortBy",
              "default": "creado_en"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TicketRead"
                  },
                  "title": "Response List Tickets Api Tickets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tickets",
        "tags": [
          "tickets"
        ]
      },
      "post": {
        "operationId": "create_ticket_api_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ticket",
        "tags": [
          "tickets"
        ]
      }
    },
...
    "/api/webhooks/tickets": {
      "post": {
        "operationId": "ingest_ticket_webhook_api_webhooks_tickets_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Deskly-Timestamp",
            "required": true,
            "schema": {
              "title": "X-Deskly-Timestamp",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Deskly-Signature",
            "required": true,
            "schema": {
              "title": "X-Deskly-Signature",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookTicketIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Ticket Webhook",
        "tags": [
          "webhooks"
        ]
      }
    }
  }
}